+++ plugins/hooks/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @zerodev/hooks
 
+## 5.2.2
+
+### Patch Changes
+
+-   Fixed contract issue
+
 ## 5.2.1
 
 ### Patch Changes