+++ plugins/permission/CHANGELOG.md
@@ -6,8 +6,14 @@
 
 - Integrated Kernel v3.1 and added kernelVersion param in account and plugins interface
 
+## 5.2.10
+
+### Patch Changes
+
+-   Decoupled creating passkey itself and the signer
+
 ## 5.2.2
 
 ### Patch Changes
 
-- Update permissionless@0.1.18
+-   Update permissionless@0.1.18