+++ wallet/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @zerodev/wallet
 
+## 0.0.9
+
+### Patch Changes
+
+- Add support for entrypoint v0.6 to KernelEIP1193Provider
+
 ## 0.0.8
 
 ### Patch Changes